// src/app/entities/kits-interface.ts
export interface KitsInterface {
    id?: number;
    id_organo?: number;
    id_categoria?: number;
    id_mantenimiento?: number;
    nombre_organo?: string;
    nombre_mantenimiento?: string;
    nombre_categoria?: string;
    id_estado?: number;
    nombre_estado?: string;
    xml_kits?: string;
}